'9:57'

+++ basic_app/models.py
@@ -3,7 +3,7 @@ from django.db import models
 
 # Create your models here.
 class Contact(models.Model):
-    phone = models.IntegerField()
+    phone = models.BigIntegerField()
 
     def __str__(self):
         return f"raqam : {self.phone}"
@@ -11,7 +11,7 @@ class Contact(models.Model):
 
 class Boglanish(models.Model):
     name = models.CharField(max_length=300)
-    phone = models.IntegerField()
+    phone = models.BigIntegerField()
 
     def __str__(self):
         return f" {self.name}, {self.phone}"
